Metadata test hygiene: semantic consolidation, doctest conversions, tightened assertions
- Move _run_ecm_scenario and every extract_code_metadata stage-scenario
  test from tests/test_inprocess_pipeline.py into
  tests/test_extract_code_metadata.py; rewrite the older hand-rolled ECM
  stage tests onto the shared helper.
- Fold tests/test_extract_code_metadata_race.py (#51 reducer-concurrency
  tests) into the same semantic file, keeping the race framing.
- Convert API-level pytest tests to doctests on the functions they test:
  extract_metadata _match_on validation errors + multi-column keys (also
  un-skipping the missing-column example), MessyConfig.needed_source_columns
  transform/format-time planning (#66/#67), MessyConfig.parse
  sources-stripped-before-resolve, and shard_subjects external-splits
  coverage + list conversion.
- Delete test_resolve_source_files_ambiguity_errors: fully subsumed by the
  pre-existing resolve_source_files ambiguity doctest in io.py.
- Tighten loose assertions (exact descriptions instead of membership
  or-chains, exact separator-joined strings, exact code_template values,
  distinct non-null hashed subject IDs) and drop stale framing (dead
  dftly-repr comment, stale helper references, source line-number section
  headers).
